+++ fastapi_viewsets/middleware/auth/authorization.py
@@ -3,8 +3,7 @@ import inspect
 from collections.abc import Awaitable, Callable
 from typing import Any, TYPE_CHECKING
 
-from fastapi import HTTPException
-
+from ...exceptions import NotAuthorizedError
 from .. import Middleware, ViewSetResult
 
 if TYPE_CHECKING:
@@ -58,7 +57,7 @@ class Authorization(Middleware):
             if inspect.isawaitable(allowed):
                 allowed = await allowed
             if not allowed:
-                raise HTTPException(status_code=403, detail="Not authorized to perform this action")
+                raise NotAuthorizedError()
 
     async def __call__(
         self,